Move chiquito circuit to right file
