+++ zkevm-circuits/src/bytecode_circuit.rs
@@ -11,9 +11,6 @@ mod test;
 #[cfg(any(feature = "test", test, feature = "test-circuits"))]
 pub use dev::BytecodeCircuit as TestBytecodeCircuit;
 
-/// chiquito version of circuit
-pub mod ccircuit;
-
 /// Bytecode circuit tester
 #[cfg(any(feature = "test", test))]
 pub mod dev;
